requirements.txt
aio-pika==9.5.5
aiohttp==3.8.1
aiormq==6.8.1
aiosignal==1.2.0
alembic==1.7.6.dev0
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.1.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chess==1.11.2
click==8.1.8
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.6
djangorestframework==3.15.2
dnspython==2.7.0
dronekit @ git+https://github.com/dronekit/dronekit-python.git@243ce0ac9c51504fc846e4ea1e35dbff03985871
duplicity==0.8.21
ecdsa==0.19.1
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
email_validator==2.2.0
empy==3.3.4
exceptiongroup==1.1.3
Faker==20.1.0
fastapi==0.115.11
fasteners==0.14.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozenlist==1.2.0
fsspec==2025.3.2
future==0.18.2
fuzzywuzzy==0.18.0
gensim==4.3.1
gevent==24.11.1
geventhttpclient==2.3.3
git-filter-repo==2.47.0
gpg==1.16.0
greenlet==3.1.1
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httpie==3.2.2
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.21.1
libcomps==0.1.15
lockfile==0.12.2
locust==2.33.2
louis==3.20.0
lxml==5.3.1
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
MAVProxy==1.8.71
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
netifaces==0.11.0
networkx==3.4.2
nltk==3.7
numpy==1.25.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pamqp==3.3.0
pandas==2.2.2
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.5.0
propcache==0.3.0
protobuf==3.12.4
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.3
proton-vpn-gtk-app==4.9.5
proton-vpn-network-manager==0.12.13
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
py==1.10.0
pyahocorasick==2.0.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymavlink==2.4.43
PyMySQL==1.1.0
PyNaCl==1.5.0
pynmeagps==1.0.46
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-apt==2.4.0+ubuntu4
python-chess==1.999
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-jose==3.4.0
python-Levenshtein==0.21.1
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.3.0
rapidfuzz==3.2.0
redis==5.2.1
regex==2021.11.10
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.0
rpm==4.17.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.11.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentence-transformers==4.0.2
sentry-sdk==1.4.3
simplejson==3.17.6
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.46.1
stockfish==3.28.0
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
tqdm==4.57.0
transformers==4.50.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
uvicorn==0.34.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==15.0.1
Werkzeug==3.1.3
Whoosh==2.7.4
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
zope.event==5.0
zope.interface==7.2
